requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
plotly>=5.18.0
